refactor(shared): name regex escaping module
