+++ src/shared/regex-escaping.ts
