+++ src/shared/text-search.ts
@@ -13,7 +13,7 @@
 import { posix, win32 } from 'node:path'
 import { assertJsonTextStructureWithinLimits } from './json-text-structure-limit'
 import { normalizeSearchResult } from './search-match-count'
-import { escapeRegex } from './string-utils'
+import { escapeRegex } from './regex-escaping'
 import type { SearchFileResult, SearchMatch, SearchOptions, SearchResult } from './types'
 
 export type SearchAccumulator = {
